CosmosDbInitializer を共通化

+++ Api/Utils/CosmosDbInitializer.cs
@@ -1,14 +1,14 @@
 ﻿using Microsoft.Azure.Cosmos.Fluent;
 using Microsoft.Azure.Cosmos;
 
-namespace Test.Api
+namespace Api.Utils
 {
 	public class CosmosDbInitializer
 	{
 		private readonly Database _database;
 
 		/// <summary>
-		/// データベースを作成します
+		/// データベースを取得または作成して、コンテナ取得準備をします
 		/// </summary>
 		/// <param name="connectionString">DB接続文字列</param>
 		/// <param name="databaseId">データベース名</param>
@@ -23,18 +23,20 @@ namespace Test.Api
 		}
 
 		/// <summary>
-		/// 空のコンテナを作成します
+		/// コンテナを取得または作成します
 		/// </summary>
 		/// <param name="containerId">コンテナID</param>
-		/// <param name="partitionKeyPath">パーティションキー</param>
-		/// <returns></returns>
-		public async Task<Container> CreateCleanContainer(string containerId, string partitionKeyPath) {
+		/// <param name="partitionKeyPath">/から始まるパーティションキー</param>
+		/// <param name="isCleanUp">初期化する？</param>
+		/// <returns>コンテナ</returns>
+		public async Task<Container> GetContainerAsync(string containerId, string partitionKeyPath, bool isCleanUp = false) {
 			var containerResponse = await _database.CreateContainerIfNotExistsAsync(containerId, partitionKeyPath);
-			if (containerResponse.StatusCode != System.Net.HttpStatusCode.Created) {
+			if (isCleanUp && containerResponse.StatusCode != System.Net.HttpStatusCode.Created) {
 				await containerResponse.Container.DeleteContainerAsync();
 				containerResponse = await _database.CreateContainerAsync(containerId, partitionKeyPath);
 			}
 			return containerResponse.Container;
 		}
+
 	}
 }
